Cloud Computing/KaizenServer/src/user/jwt_handler.py
import jwt
from datetime import datetime, timedelta, timezone
from typing import Dict

SECRET_KEY = "simuhamuhimuha"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

def create_access_token(data: dict) -> str:
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

def verify_token(token: str) -> dict:
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        print(f"Decoded JWT payload: {payload}")
        if payload.get("exp") >= datetime.now(timezone.utc).timestamp():
            return payload
        return None
    except jwt.ExpiredSignatureError:
        return None
    except jwt.InvalidTokenError:
        return None

def token_response(token: str) -> Dict[str, str]:
    return {
        "access_token": token,
        "token_type": "bearer"
    }

def signJWT(user_id: str) -> Dict[str, str]:
    payload = {
        "user_id": user_id,
        "exp": datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    }
    token = create_access_token(payload)
    return token_response(token)

def decodeJWT(token: str) -> dict:
    print(token);
    return verify_token(token)
