Cloud Computing/KaizenServer/requirements.txt
fastapi
uvicorn
alembic
SQLAlchemy
PyJWT
python-dotenv
psycopg2-binary
pydantic
passlib
bcrypt